Split broadcaster-specific code into its own file.
